requirements.txt
flask
requests
pandas
matplotlib
selenium
undetected-chromedriver
pdfplumber
python-dotenv
playwright
